2010-03-22 running of model separated in running by age-sex group if number of simulated persons is large, as old version was not able to run enough persons without getting out of memory  error.

+++ Dynamo-HIA_UI/src/nl/rivm/emi/dynamo/estimation/SimulationConfigurationFactory.java
@@ -286,7 +286,7 @@ public class SimulationConfigurationFactory {
 				// random seeds meegeven: dat gaat nu per persoon dus niet
 				// meer nodig
 				writeFinalElementToDom(rootElement, "randomSeed",
-						((Integer) scenInfo.getRandomSeed()).toString());
+						((Long) scenInfo.getRandomSeed()).toString());
 				document.appendChild(rootElement);
 				writeDomToXML(ConfigXMLfileName, document);
